fix: adversarial review round 2 — 7 more bugs fixed
Dead code removed:
- shaders/generic/per_head_rms_norm.wgsl (zero callers; latent WG_SIZE mismatch)
- shaders/generic/fused_norm_add.wgsl (zero callers; optimization never wired up)

Real bugs fixed:
- gguf_loader.py: BF16→F16 cast silently produced ±inf for weights with
  |value| > 65504 (F16_MAX < BF16_MAX). Added np.clip before the cast.
- gguf_loader.py: GGUF weight buffers missing COPY_SRC flag; to_numpy()
  would raise GPUValidationError on any weight buffer read. Added COPY_SRC.
- llama.py, gemma4.py: assert for multi-sequence guard removed by python -O;
  replaced with explicit RuntimeError.
- model_runner.py: head_dim computed as hidden_size // num_attention_heads,
  ignoring Gemma4's explicit head_dim config field. Used getattr fallback.
- pipeline.py: FileNotFoundError from read_text() gave no indication of
  which shader was missing; added exists() check with diagnostic message.

Test precision fix:
- test_kernels_attn.py: v_vectors.mean() accumulated in f16 (numpy default)
  but the shader accumulates in f32; switched to f32 mean to match.

+++ tests/test_kernels_attn.py
@@ -430,7 +430,9 @@ def test_attention_equal_keys_uniform_weights(wgpu_device):
 
     # Invariant: identical K vectors → uniform attention weights → output = mean(V)
     # Both Q-heads use the same KV cache (single KV head, GQA ratio=2)
-    v_mean = v_vectors.mean(axis=0).astype(np.float16)  # mean of 4 V vectors
+    # Compute reference mean in float32 to match shader's f32 accumulation precision.
+    # v_vectors.mean() in f16 accumulates in f16, diverging from the shader's f32 path.
+    v_mean = v_vectors.astype(np.float32).mean(axis=0).astype(np.float16)
     for q_head in range(num_q_heads):
         np.testing.assert_allclose(
             result[q_head].astype(np.float32), v_mean.astype(np.float32),
